requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
ecdsa==0.19.1
idna==3.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.0
urllib3==2.2.3
